Viser indikator når agent skriver.
