+++ nwb.demo.config.js
@@ -8,7 +8,7 @@ module.exports = {
         extensions: ['.ts', '.tsx', '.js', '.jsx']
       },
       module: {
-        rules: [{ test: /\.tsx$/, loader: 'ts-loader' }]
+        rules: [{ test:  /\.ts|\.tsx$/, loader: 'ts-loader' }]
       }
     },
     rules: {
